pulls getProgress into it's own file to deal with circular dependency between pinnedStudent and users model
